Sound from MCP4822 SPI DAC

+++ myringbuffer.h
@@ -14,8 +14,8 @@ extern "C" {
 #define MYRINGBUFFER_MASK (MYRINGBUFFER_SIZE - 1)
 
 void     my_rb_init(void);
-bool     my_rb_put(uint8_t sample);
-bool     my_rb_get(uint8_t *sample);
+bool     my_rb_put(uint16_t sample);
+bool     my_rb_get(uint16_t *sample);
 uint32_t my_rb_free(void);
 uint32_t my_rb_used(void);
